refactor: Update MetricType enum with Top_Five_Daily option
Update the MetricType enum in the MetricType.ts file to include a new option called Top_Five_Daily. This change allows for more specific tracking of metrics in the application.

Update chart.js package version to 4.4.5

In the package.json file, update the chart.js package version to 4.4.5. This update includes bug fixes and improvements to the charting functionality in the application.

Update vue-chartjs package version to 5.3.1

In the package.json file, update the vue-chartjs package version to 5.3.1. This update includes bug fixes and improvements to the charting functionality in the application.

Update common.js in en and es locales

In the common.js files for the en and es locales, add new translations for weekly, daily, daily_top_5, weekly_top_5, year_service_progress, and first_place. These translations are used in various parts of the application to display text to the user.

Update Dashboard.vue and Metrics.vue components

In the Dashboard.vue and Metrics.vue components, make various updates and improvements. These updates include importing the useMetricsStore and useDriversStore from their respective stores, and adding a new method getCurrentYearMetric to the useMetricsStore. These changes improve the functionality and performance of the components.

Add MetricRepository.ts file

Add a new file called MetricRepository.ts to the repositories directory. This file contains the MetricRepository class, which provides methods for retrieving metrics data from the database.

Update Metrics.vue component

In the Metrics.vue component, make various updates and improvements. These updates include importing the useMetricsStore and useI18n from their respective stores, and adding a new method getTop5DailyMetric to the useMetricsStore. These changes improve the functionality and performance of the component.

Update Metrics.vue component

In the Metrics.vue component, update the chartOptions and dailyTop5chartOptions variables to include new options for the charts. These options improve the appearance and functionality of the charts in the component.

Update Metrics.vue component

In the Metrics.vue component, update the onBeforeMount hook to call the getTop5DailyMetric method from the useMetricsStore. This change ensures that the top 5 daily metrics are retrieved and displayed correctly in the component.

Update Metrics.vue component

In the Metrics.vue component, update the dailyTop5ChartData variable to include the data for the top 5 daily metrics. This change ensures that the data is correctly displayed in the chart.

Update Metrics.vue component

In the Metrics.vue component, update the template to include new card elements for displaying the year service progress and cancel percent metrics. These changes improve the layout and organization of the component.

Update Metrics.vue component

In the Metrics.vue component, update the template to include a select element for choosing the frequency of the top 5 metrics (daily or weekly). This change allows the user to switch between daily and weekly metrics in the chart.

Update Metrics.vue component

In the Metrics.vue component, update the template to include a new chart element for displaying the top 5 daily metrics. This change adds a new chart to the component for visualizing the top 5 daily metrics.

Update Metrics.vue component

In the Metrics.vue component, update the template to include loading spinners for the global, percent, and top 5 daily charts. These spinners are displayed while the data is being loaded and improve the user experience.

Update MetricsStore.ts file

In the MetricsStore.ts file, add a new method getTop5DailyMetric to the useMetricsStore. This method retrieves the top 5 daily metrics from the MetricRepository and updates the top5DailyMetric map in the store.

Add MetricRepository.ts file

Add a new file called MetricRepository.ts to the repositories directory. This file contains the MetricRepository class, which provides methods for retrieving metrics data from the database.

Update MetricsStore.ts file

In the MetricsStore.ts file, update the getDailyTopFive method to use the MetricRepository to retrieve the top 5 daily metrics. This change improves the performance and maintainability of the method.

Update MetricsStore.ts file

In the MetricsStore.ts file, add a new map called top5DailyMetric to the store. This map stores the top 5 daily metrics for each driver in the application.

Update MetricsStore.ts file

In the MetricsStore.ts file, update the getTop5DailyMetric method to use the MetricRepository to retrieve the top 5 daily metrics. This change ensures that the top 5 daily metrics are correctly retrieved and stored in the top5DailyMetric map.

Update MetricsStore.ts file

In the MetricsStore.ts file, update the getTop5DailyMetric method to use the findById method from the useDriversStore. This change ensures that the driver information is correctly retrieved and used in the top 5 daily metrics.

Update Metrics.vue component

In the Metrics.vue component, update the onBeforeMount hook to call the getTop5DailyMetric method from the useMetricsStore. This change ensures that the top 5 daily metrics are retrieved and displayed correctly in the component.

Update Metrics.vue component

In the Metrics.vue component, update the dailyTop5ChartData variable to include the data for the top 5 daily metrics. This change ensures that the data is correctly displayed in the chart.